title-deed-application/src/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
area==1.1.1
async-timeout==4.0.3
attrs==23.1.0
bitarray==2.8.2
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
cytoolz==0.12.2
DateTime==5.2
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.1
eth-utils==2.3.0
fiona==1.9.5
Flask==3.0.0
fonttools==4.44.0
frozenlist==1.4.0
hexbytes==0.3.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lru-dict==1.3.0
MarkupSafe==2.1.3
matplotlib==3.8.1
multidict==6.0.4
numpy==1.26.1
packaging==23.2
pandas==2.1.2
parsimonious==0.9.0
Pillow==10.1.0
protobuf==4.25.0
py-solc-x==1.1.1
pycryptodome==3.19.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyunormalize==15.0.0
pywin32==306
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.12.0
semantic-version==2.10.0
shapely==2.0.2
six==1.16.0
toolz==0.12.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
web3==6.11.2
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.2
zipp==3.17.0
zope.interface==6.1
